% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_standprop.R
\name{approx_standprop}
\alias{approx_standprop}
\title{Create a daily sequence from yearly values using interpolation}
\usage{
approx_standprop(x.years, y, xout.years = x.years, approx.method)
}
\arguments{
\item{x.years}{A sequence of years or a single year}

\item{y}{vector of the same length as x.years}

\item{xout.years}{years for which ouput is generated, maybe longer or shorter than
x.years}

\item{approx.method}{name of interpolation method 'constant' or 'linear'.}
}
\value{
A vector of daily values corresponding to the dates in years.out
}
\description{
Uses yearly values of inter-annual vegetation development values (maxlai, sai, height, densef, age)
and interpolates them to a daily sequence.
}
\details{
The y values are interpreted to be valid at the 1st of January of the respective x.years.
Interpolation is made with rule = 2, so that for daily values outside the interpolation interval
the value at the closest data extreme is used. For \code{approx.method = 'linear'}, this results in
constant values for all dates within \code{xout.years >= max(x.years)} or \code{xout.years < min(x.years)}.
If a single value is passed via y, then no interpolation is made and the value is simply repeated with
length of Dates within xout.years.
}
