% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptf_hamken_ff.R, R/ptf_wessolek_mvg.R
\name{hydpar_wessolek_mvg}
\alias{hydpar_wessolek_mvg}
\title{Mualem - van Genuchten Parameters for forest floor organic soil horizons, from Hammel & Kennel 2001}
\usage{
hydpar_wessolek_mvg(tex.KA5)

hydpar_wessolek_mvg(tex.KA5)
}
\arguments{
\item{tex.KA5}{A character vector with soil texture values according to KA5}

\item{n}{An integer value specifying the number of rows the returned data.frame should have
(i.e. the number of repetitions of the MvG-Parameter set)}
}
\value{
A data.frame with the number of rows equal to length(tex.KA5)
The data.frame contains the following variables:
\describe{
  \item{ths}{Saturation water content fraction}
  \item{thr}{Residual water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, hPa-1}
  \item{n}{N parameter of van Genuchten water retention function}
  \item{m}{M parameter of van Genuchten water retention function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
}

A data.frame with the number of rows equal to length(tex.KA5)
The data.frame contains the following variables:
\describe{
  \item{tex.KA5}{soil texture according to KA5 }
  \item{thr}{Residual water content fraction}
  \item{ths}{Saturation water content fraction}
  \item{alpha}{Alpha parameter of van Genuchten water retention function, hPa-1}
  \item{n}{N parameter of van Genuchten water retention parameter}
  \item{tort}{Tortuosity parameter of Mualem hydraulic conductivity function}
  \item{ksat}{Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function, mm d-1}
}
}
\description{
Derives a set of values for MVG Parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
for forest floor horizons as decribed in Hammel & Kennel 2001.

Derives the MVG Parameters ThetaR, ThetaS, Alpha, n, Tort and Ksat
    from KA5 soil texture classes
}
\references{
Hammel & Kennel (2001) Charakterisierung und Analyse der Wasserverfügbarkeit
                und des Wasserhaushalts von Waldstandorten in Bayern mit dem Simulationsmodell BROOK90,
                Forstliche Forschungsberichte München 185

Wessolek, Kaupenjohann and Renger (2009) Bodenphysikalische Kennwerte
                und Berechnungsverfahren f?r die Praxis. Boden?kologie und Bodengenese 40,
                Berlin, Germany (Table No. 10)
}

