% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_IniControl.R
\name{MakeIniControl.B90}
\alias{MakeIniControl.B90}
\title{Create a list of model options}
\usage{
MakeIniControl.B90(...)
}
\arguments{
\item{...}{named arguments to be included in return value}
}
\value{
returns a list of default model options to play with. Use to run LWF-Brook90.
}
\description{
Create a list of model options
}
\details{
\describe{
  \item{startdate}{Startdate of simulation.}
  \item{enddate}{Enddate of simulation.}
  \item{fornetrad}{Use global solar radiation (="globrad") or sunshine duration hours (="sunhour") for net radiation calculation?}
  \item{precinterval}{Number of precipitation intervals (default is 1). If prec.interval > 1, a separate file has to be provided! }
  \item{coords_x}{Longitude value (decimal degrees) of the simulation location (without effect).}
  \item{coords_y}{Latitude value (decimal degrees) of the simulation location.}
  \item{budburst}{Calculate budburst day of year dynamically (="dynamic") or use fixed values (="fixed") defined in parameters?}
  \item{budburst.method}{Name of method for dynamic budburst calculation. Passed to 'start.method'-argument of  \code{\link[vegperiod]{vegperiod}}.}
  \item{leaffall}{ Calculate leaffall day of year dynamically (="dynamic") or use fixed values (="fixed") defined in parameters?}
  \item{species_dynbudburst}{name of tree species [required if budburst.method='Menzel']. Passed to 'species'-argument of \code{\link[vegperiod]{vegperiod}} .}
  \item{longtermdyn}{Name of method for longterm plant development: either constant ("const") or yearly changing values ("table") of plant development.}
  \item{annuallaidyn}{Name of method method for generating seasonal plant development. Passed to 'method'-argument of \code{\link{MakeSeasLAI}}. }
  \item{imodel}{Name of hydraulic parameterization: "CH" for Clapp/Hornberger, "MvG" for Mualem/van Genuchten}
  \item{rootmodel}{Model name of the root length density depth distribution function. Any of the names accepted by \code{\link{MakeRelRootDens}} are allowed.
  Assign "soilvar", if the root length density should be taken from the soil-data.frame (colname relrootlength)     }
  \item{humusroots}{Shall roots be defined for humuslayer? If TRUE, then humus layer will have the same root length density as the uppermost mineral soil layer.}
}
}

