% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeOptions.R
\name{MakeOptions.B90}
\alias{MakeOptions.B90}
\title{Create a list of model control options}
\usage{
MakeOptions.B90(...)
}
\arguments{
\item{...}{named arguments to be included in return value}
}
\value{
returns a list of default model options. Can be used as inicontrol-argument in \code{\link[Run.B90]{Run.B90}}.
}
\description{
Create a list of model control options
}
\details{
\describe{
  \item{startdate}{Startdate of the simulation.}
  \item{enddate}{Enddate of the simulation.}
  \item{fornetrad}{Use global solar radiation (="globrad") or sunshine duration
  hours (="sunhour") for net radiation calculation?}
  \item{prec.interval}{Number of precipitation intervals (default is 1).
  If prec.interval > 1, a separate file ("in/PRFILE.DAT") has to be provided manually!}
  \item{richter.prec.corr}{ Correct Precipitation for wind and evaporation losses (not implented yet)}
  \item{budburst}{Calculate budburst day of year dynamically (="dynamic") or use fixed values (="fixed") defined in parameters (budburstdoy)?}
  \item{budburst.method}{Name of method for budburst calculation. If
  'constant' or 'fixed', budburst day of year from parameters is used.
  All other methods calculate budburst day of year dynamically from temperatures, and
  the method name is passed to the 'start.method'-argument of  \code{\link[vegperiod]{vegperiod}}.}
  \item{leaffall.method}{Name of method for leaffall calculation. If
  'constant' or 'fixed', beginning of leaffall (day of year) from parameters is used.
  All other methods calculate budburst day of year dynamically from temperatures, and
  the method name is passed to the 'end.method'-argument of  \code{\link[vegperiod]{vegperiod}}.}
  \item{longtermdyn}{Name of method for longterm plant development: either constant ("constant")
  or yearly values ("table") of plant development.}
  \item{lai.method}{Name of method method for generating daily plant development.
  Passed to 'method'-argument of \code{\link{MakeSeasLAI}}. }
  \item{imodel}{Name of hydraulic parameterization: "CH" for Clapp/Hornberger, "MvG" for Mualem/van Genuchten}
  \item{rootmodel}{Model name of the root length density depth distribution function. Any of the names accepted by \code{\link{MakeRelRootDens}} are allowed.
  Assign "soilvar", if the root length density should be taken from the soil-data.frame (colname relrootlength)     }
  \item{humusroots}{Shall roots be defined for humuslayer? If TRUE, the humus layers will have the same root length density as the uppermost mineral soil layer.}
}
}
\examples{
# Default options
options.b90 <- MakeOptions.B90()
# Include specific options
options.B90_dynamic_phenology <- MakeOptions.B90(budburst = 'Menzel', leaffall ='vonWilpert')
}
