% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeRoots.R
\name{MakeRelRootDens}
\alias{MakeRelRootDens}
\title{Generates a root density depth function for the soil nodes}
\usage{
MakeRelRootDens(soilnodes, maxrootdepth, method, param, relrootden = NULL,
  rootdepths = NULL, humusroots = FALSE)
}
\arguments{
\item{soilnodes}{vector of lower soil layer limits (cm, positive downward.),
for which the relative root distribution will be calculated. Values in  Topsoil
(humus) layers are given as negative values.}

\item{maxrootdepth}{the maximum rooting depth below  which relative root
density length will be set to zero}

\item{method}{root distribution model name. Choose "betamodel" to use the model after Gale & Grigal, "table" to supply a table,
or "constant" to get a uniform root distribution with depth}

\item{param}{parameter(s) of the root distribution function}

\item{relrootden}{vector with relative root length}

\item{rootdepths}{vector with lower depths limit, corresponding to relative
rootlength-vector}

\item{humusroots}{Should there be roots in top soil layers? If TRUE, the relative
rootlength of the topsoil layers will be the same as in the upper mineral soil layer.}
}
\value{
vector of relative rootlength, corresponding to soilnodes
}
\description{
Generates a root density depth function for the soil nodes
}
