% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStand.R
\name{MakeStand}
\alias{MakeStand}
\title{Create a daily sequence of stand properties (lai, sai, height, age)}
\usage{
MakeStand(stand.years, sai, height, densef, age, years.out = stand.years,
  approx.method)
}
\arguments{
\item{stand.years}{A sequence of years or a single year, at which the stand properties were measured}

\item{sai}{Stem area index, either a fixed single value or corresponding to
stand.years.}

\item{height}{Vegetation height, either a fixed single value or corresponding
to stand.years. m}

\item{densef}{Relative vegetation density values, either a fixed single
value orcorresponding to stand.years.}

\item{age}{Vegetation age, internally used in LWF-Brook90 for root growth (see parameters
rgrorate, rgroper,inirtlen, inirdep). Either a fixed single value or corresponding
to the beginning of stand.years, or corresponding to stand.years. a}

\item{years.out}{years for which ouput is generated, maybe longer or shorter than
stand.years,but must include at least one of the values in stand.year.}

\item{maxlai}{Maximum leaf area index during summer, either a fixed single value or
corresponding to stand.years}

\item{winlaifrac}{Minimum leaf area index during winter, either a fixed single value or
corresponding to stand.years.}

\item{budburst.doy}{Day of year when leaf flushing sets in, either a fixed single
value or corresponding to stand.years.}

\item{leaffall.doy}{Day of year when leaffall begins, either a fixed single value
or corresponding to stand.years.}

\item{...}{parameters passed to \code{\link{MakeSeasLAI}}.}
}
\value{
Returns a data.frame with columns lai, sai, height,
densef at daily resolution
}
\description{
Uses a table with yearly values of perennial vegetation development or anually constant
values (maxlai, winlaifrac, sai, height, age) and turns them into a daily table of vegetation
properties, that can be passed on to write.climate.in
}
\details{
Linear interpolation is used to  generate daily sequences from yearly values of
sai, height, densef and age. Daily leaf area index is generated by \code{\link{MakeSeasLAI}}.
}
