% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRunBrook90.R
\name{MultiRun.B90}
\alias{MultiRun.B90}
\title{Make a multirun simulation using a set of variable input parameters.}
\usage{
MultiRun.B90(nRuns, param_fixed, param_var, multirun.dir = "MultiRuns",
  keep.outfiles = FALSE, singlerun_names = paste0("RunNo", 1:nRuns),
  cores = 3, showProgress = TRUE, ...)
}
\arguments{
\item{nRuns}{Number of runs}

\item{param_fixed}{Named list of model constant input parameters to use in all single runs.}

\item{param_var}{Named list or data.frame of variable input parameters with realisations for each single run.}

\item{multirun.dir}{The directory where to write temporary files of single runs}

\item{singlerun_names}{character vector with the names of the single runs.}

\item{cores}{Number of cores to use for parallel processing}

\item{...}{Arguments passed to \code{\link[Run.B90]{Run.B90}}.
At least inicontrol, climate, and soil have to be provided!}
}
\description{
Make a multirun simulation using a set of variable input parameters.
}
