% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retentioncurve_MvG.R
\name{MvG.swc}
\alias{MvG.swc}
\title{Retention-Curve: water content vs. pressure head, van Genuchten}
\usage{
MvG.swc(psi, alpha, n, ThS, ThR, m = 1 - 1/n)
}
\arguments{
\item{psi}{vector of pressure heads}

\item{alpha}{value vG-alpha, unit: reciprocal pressure head}

\item{n}{value of van Genuchten parameter N}

\item{ThS}{n value of van Genuchten parameter Ths (saturation water content)}

\item{ThR}{n value of van Genuchten parameter Thr residual water content}

\item{m}{value of van Genuchten parameter N. Defaults 1-1/n}
}
\value{
vector of water contents
}
\description{
Retention-Curve: water content vs. pressure head, van Genuchten
}
