% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rootden_slb1}
\alias{rootden_slb1}
\title{Fine root distribution at Solling Beech experimental site 'SLB1'}
\format{A data.table with 9 rows and 2 variables
\describe{
  \item{depth}{lower depth of the root layer}
  \item{rootden}{root density (number of roots per decimeter soil depth)}
}}
\usage{
rootden_slb1
}
\description{
A dataset containing the number of fine roots per decimeter soil depth
}
\keyword{datasets}
