% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunBrook90.R
\name{Run.B90}
\alias{Run.B90}
\title{Run the LWF-Brook90 hydrological model and return results}
\usage{
Run.B90(directory, inicontrol, param, climate, soil, longtermdev,
  outputmat = choose_output.B90(edit = FALSE), write.climate.in = TRUE,
  write.param.in = TRUE, output.plant.devt = TRUE, output_log = "",
  keepoutputfiles = TRUE, verbose = TRUE, path_b90.exe = "b90.exe",
  run.model = TRUE)
}
\arguments{
\item{directory}{Directoryname of the project where the input and output files
are located. Will be created, if not existing.}

\item{inicontrol}{Named list of model control options. Use
\code{\link{MakeInicontrol.B90}} to generate a list with default model control options.}

\item{param}{Named list of model input parameters. Use
\code{\link{MakeParam.B90}} to generate a list with default model parameters.}

\item{climate}{Data.frame with daily climate data. All names(climate) should be
found in and corrrespond to arguments dates, tmax, tmin, wind, prec, vappres,
globrad, sunhours) of \code{\link{writeClimate.in}}.}

\item{soil}{Data.frame containing the hydraulic properties of the soil layers.
Each row represents one layer, containing the layer's boundaries
['upper' and 'lower' negative downwards, m], van Genuchten retention parameters 'alpha' [m-1], 'n', 'ths',
'thr', Mualem conductivity function parameters ksat [mm d-1] and t, and the volume fraction
of gravel content.}

\item{longtermdev}{Data.frame with yearly values of vegetation properties
(years, age, lai, sai, height, densef) that are passed to \code{\link{MakeStand}}.}

\item{outputmat}{A [10,5]-matrix marking the desired model-output. Use
\code{\link{Choose_output.B90}} to generate and edit it.}

\item{write.climate.in}{Should the climate file be written? Ignored if no
Climate.in is found in directory, Climate.in is written anyway.}

\item{write.param.in}{Should the climate file be written? Ignored if
write.climate.in == TRUE and written anyway.}

\item{output.plant.devt}{Return daily values of aboveground plant properties?}

\item{output_log}{Logical or name where stdout from \code{\link[base]{system2}}-call
is sent. The default output_log = FALSE discards stdout and slightly reduces execution
time of LWF Brook90. output_log = "" sents all command-line model-output to the R
console, all other names will create a text file containing the model runtime output.}

\item{keepoutputfiles}{Keep the model output files after running and
returning output? Default is TRUE.}

\item{verbose}{Print messages to screen? Default is TRUE.}

\item{path_b90.exe}{Filename of the executable. The default setting looks for the
executable 'b90.exe' within 'directory'.}

\item{run.model}{Run the model or only create input files? Default is TRUE.}
}
\value{
Returns a list of data.frames (data.tables) containing the model results
and some additional informations including the execution time.
}
\description{
Takes all necessary information needed to run the LWF-Brook90 hydrological model,
writes input files, starts the external executable via a system-call and returns
the results.
}
\examples{

#Set up lists containing model control options and model parameters:

param.b90 <- MakeParam.B90()
options.b90 <- MakeIniControl.B90()

Set start and end Dates for the simulation

options.b90$startdate <- as.Date("2000-01-01")
options.b90$enddate <- as.Date("2004-12-31")

# Derive soil hydraulic properties from soil physical properties
# using pedotransfer functions

soil <- cbind(soil_slb1[-c(1:3),], hydpar_wessolek_mvg(soil_slb1$texture[-c(1:3)]))

Run LWF-Brook90

names(soil) <- tolower(names(soil))
names(soil)[c(14,15)] <- c("npar", "mpar")
b90.result <- Run.B90(directory = "example_run_b90",
                      inicontrol = options.b90,
                      param = param.b90,
                      climate = meteo_slb1,
                      soil = soil,
                      path_b90.exe = "H:/B90/MultiControl/b90.exe")
}

