% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_slb1}
\alias{soil_slb1}
\title{Soil profile data from the Solling Beech experimental site 'SLB1'}
\format{A data.frame with 20 rows and 11 variables
\describe{
  \item{horizon}{horizon symbol}
  \item{upper}{upper layer boundary, m}
  \item{lower}{lower layer boundary, m}
  \item{texture}{soil texture according to German soil texture classification system}
  \item{bd}{bulk density of the fine earth, g cm-3}
  \item{gravel}{fraction of coarse material }
  \item{sand}{sand content, mass-\% }
  \item{silt}{silt content, mass-\% }
  \item{clay}{clay content, mass-\% }
  \item{c_org}{organic carbon content, mass-\%}
}}
\usage{
soil_slb1
}
\description{
A dataset containing the soil horizons' physical properties
}
\keyword{datasets}

