% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeclimate.R
\name{writeClimate.in}
\alias{writeClimate.in}
\title{Writes the LWF-Brook90 input file Climate.in}
\usage{
writeClimate.in(dates, tmax, tmin, vappres, wind, prec, globrad, sunhours,
  densef, height, lai, sai, age, latitude, mesfl = rep(0, length(dates)),
  filename = "in/Climate.in", use.sunhours = F, richter.prec.corr = F,
  prec.int = 1, snow.ini = 0, gwat.ini = 0)
}
\arguments{
\item{dates}{date sequence with daily resolution.}

\item{tmax}{daily maximum air temperature, \eqn{°C}.}

\item{tmin}{daily minimum air temperature, \eqn{°C}.}

\item{vappres}{daily mean air vapor pressure, \eqn{kPa}.}

\item{wind}{daily mean wind speed, \eqn{m s^{-1}}.}

\item{prec}{daily precipitation, {mm d^{-1}}.}

\item{globrad}{solar radiation on the zero plane, \eqn{MJm^{-2}d^{-1}}}

\item{sunhours}{sunshine duration, hours per day}

\item{densef}{canopy density multiplier between 0.05 and 1, dimensionless}

\item{height}{canopy height, m}

\item{lai}{projected leaf area index, \eqn{m^{2} m^{-2}}}

\item{sai}{projected stem area index, \eqn{m^{2} m^{-2}}}

\item{age}{stand age, years}

\item{latitude}{geographic latitude of the simulated site, decimal degrees}

\item{mesfl}{measured stream flow,  \eqn{mm d^{-1}}}

\item{filename}{filename for writing. Defaults to '~in/Climate.in}

\item{use.sunhours}{Use sunshine duration hours rather than global radiation?}

\item{richter.prec.corr}{Correct precipitation values for evaporation and wind losses?}

\item{prec.int}{number of precipitation intervals per day. Currently only 1 interval per day is supported.}

\item{snow.ini}{initial amount of snow on the ground, mm water equivalent}

\item{gwat.ini}{initial groundwater storage below soil layers, mm}
}
\value{
no return value
}
\description{
Creates the Climate.in file from daily climate data and vegetation properties,
providing some flexibility with regards to radiation (solar radiation or sunshine hours)
and precipitation correction
}

