% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeParam.in.R
\name{writeParam.in}
\alias{writeParam.in}
\title{Writes the LWF-Brook90 input file Param.in}
\usage{
writeParam.in(b90opts, parameters, materials, soil, outmat, filename)
}
\arguments{
\item{b90opts}{A named list of model control options.}

\item{parameters}{A named list of model parameters.}

\item{materials}{A data.frame with soil materials holding unique hydraulic parameters.}

\item{soil}{A data.frame containing the discretisation of model soil layers.}

\item{outmat}{A [5,10] matrix marking the desired model output.}

\item{filename}{filename for writing. Defaults to '~in/Param.in'.}
}
\value{
no return value
}
\description{
Creates the Param.in file for the LWF-Brook90 executable
}

