% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeRoots.R
\name{MakeRelRootDens}
\alias{MakeRelRootDens}
\title{Generates a root density depth function for the soil nodes}
\usage{
MakeRelRootDens(soilnodes, maxrootdepth = min(soilnodes),
  method = "betamodel", beta = 0.97, relrootden = NULL,
  rootdepths = NULL, humusroots = FALSE)
}
\arguments{
\item{soilnodes}{vector of lower soil layer limits,
for which the relative root distribution will be calculated. Topsoil (humus) layers
corespond to negative values.}

\item{maxrootdepth}{the maximum rooting depth (m, positive downwards) below which
relative root length density will be set to zero}

\item{method}{root distribution model name. Choose "betamodel" to use the model after Gale & Grigal, "table" to supply a table,
"constant" to get a uniform root distribution with depth. 'table' redistributes layers of root density to the soilnodes}

\item{beta}{parameter(s) of the root distribution function}

\item{relrootden}{vector with relative root density}

\item{rootdepths}{vector with lower depths limit, corresponding to relative
rootlength-vector}

\item{humusroots}{Should there be roots in top soil layers? If TRUE, the relative
rootlength of the topsoil layers will be the same as in the upper mineral soil layer.}
}
\value{
vector of relative rootlength, corresponding to soilnodes
}
\description{
Generates a root density depth function for the soil nodes
}

