% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSeasLAI.R
\name{MakeSeasLAI}
\alias{MakeSeasLAI}
\title{Create daily seasonal plant dynamics sequence (e.g. leaf area index development)
for one year}
\usage{
MakeSeasLAI(method = "b90", year, maxlai, winlaifrac = 0,
  budburst.doy = 121, leaffall.doy = 279, emerge.dur = 28,
  leaffall.dur = 58, shape.optdoy = 220, shape.budburst = 0.5,
  shape.leaffall = 10, lai.doy = c(1, 121, 150, 280, 320, 365),
  lai.frac = c(0, 0, 0.5, 1, 0.5, 0))
}
\arguments{
\item{method}{name of method for generating the sequence. Must be one of "b90", "linear", "Coupmodel"}

\item{year}{year to identify leap years}

\item{maxlai}{maximum value during summer (not used when method = 'linear')}

\item{winlaifrac}{minimum lai as fraction of maxlai during winter (not used when method = 'linear')}

\item{budburst.doy}{day of year when growth begins (not used when method = 'linear')}

\item{leaffall.doy}{day of year when growth cessates (not used when method = 'linear')}

\item{emerge.dur}{number of days until maximum value is reached (required when method = "Coupmodel")}

\item{leaffall.dur}{number of days until minimum value is reached (required when method = "b90")}

\item{shape.budburst}{shape parameter for the growth phase (required when method = "Coupmodel")}

\item{shape.leaffall}{shape parameter growth cessation (required when method = "Coupmodel")}

\item{lai.doy}{vector of integers (required when method = "linear")}

\item{lai.frac}{vector of values corresponding to doy.values (required when method = "linear")}

\item{opt.doy}{day of year when optimum value is reached (required when method = "Coupmodel")}
}
\value{
Returns a seasonal sequence of length maxdoy
}
\description{
Creates the seasonal course of leaf area index, based
on minimum lai, maximum lai, budburst and leaffall day of year (doy) and shape
parameters.
}
