% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_to_rbrook90.R
\name{param_to_rbrook90}
\alias{param_to_rbrook90}
\title{Writes the LWF-Brook90 input file Param.in}
\usage{
param_to_rbrook90(options.b90, param.b90, n_soillays, n_materials)
}
\arguments{
\item{options.b90}{a named list of model control options.}

\item{param.b90}{a named list of model parameters.}

\item{n_soillays}{number of model soil layers.}

\item{n_materials}{number of soil materials}
}
\value{
a numerical vector holding all parameters required by PFILE.INC
in the right order.
}
\description{
Creates the Param.in file for the LWF-Brook90 executable
}
